pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
folium = "^0.17.0"
pandas = "^2.2.2"
streamlit = "^1.36.0"
streamlit-folium = "^0.21.1"
google-generativeai = "^0.7.2"
openai = "^1.35.14"
requests = "^2.32.3"
markdown = "^3.6"
beautifulsoup4 = "^4.12.3"
langchain = "^0.2.9"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"